Commissions (part 1)

+++ database-service/src/main/java/com/example/database_service/repository/WorkshopRepository.java
@@ -3,5 +3,5 @@ package com.example.database_service.repository;
 import com.example.database_service.entity.Workshop;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface WorkshopRepository extends JpaRepository<Workshop, Long> {
+public interface WorkshopRepository extends JpaRepository<Workshop, Integer> {
 }
